Add support for Python 3.
In Python 3 the C char * maps on to the Python bytes type, rather
than the str type as it does in Python 2. These changes modify
pystatgrab to convert between the two so that the library behaves
in the same way on both Python 3 and 2. The test suite passes on
both with no changes.

For future reference it's worth noting that the change on line 303
of statgrab.pyx had to be done before the for loop that follows it.
Doing it inside the for loop resulted in the data inside vs being
corrupted after the loop ended, presumably because it went out of
scope. It wasn't immediately obvious what was happening, thus the
mention in the commit log.