Add CI for pystatgrab.
This builds a pystatgrab distfile, as end users would receive, and
then tests it on different Linux operating systems. It tests against
both the bundled libstatgrab that comes with the OS, and also against
the current libstatgrab master which it builds separately. It also
tests against both python 2 and 3.

Finally, it does GitLab's security and code quality testing.